requirements.txt
lxml
urlextract
readability
pymongo
absl-py==0.4.0
astor==0.7.1
beautifulsoup4==4.6.3
bleach==2.1.4
boto==2.49.0
boto3==1.8.1
botocore==1.11.1
bz2file==0.98
certifi==2018.8.24
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.3.0
docutils==0.14
entrypoints==0.2.3
first==2.0.1
gast==0.2.0
gensim==3.5.0
grpcio==1.14.1
h5py==2.8.0
html5lib==1.0.1
idna==2.7
ipykernel==4.8.2
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
Jinja2==2.10
jmespath==0.9.3
joblib==0.12.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mecab-python3==0.7
mistune==0.8.3
mojimoji==0.0.8
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.14.5
pandas==0.23.4
pandocfilters==1.4.2
patsy==0.5.0
pexpect==4.6.0
pickleshare==0.7.4
pip-tools==2.0.2
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
pymc3==3.5
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.19.1
s3transfer==0.1.13
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
smart-open==1.6.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.2
tornado==5.1
tqdm==4.25.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.0
